requirements.txt
# Sutra-Markdown V2 - Core Dependencies

# ============================================================================
# Core Framework
# ============================================================================
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.9

# ============================================================================
# Async & Concurrency
# ============================================================================
asyncio==3.4.3
aiofiles==23.2.1
aioboto3==12.3.0
httpx>=0.25.2,<0.26.0

# ============================================================================
# Nomic Embeddings (Revolutionary Intelligence Layer)
# ============================================================================
# Note: Install from official Nomic sources
# pip install nomic
nomic>=3.0.0

# Alternative: sentence-transformers for loading Nomic models
sentence-transformers==2.3.1

# ============================================================================
# Document Parsing
# ============================================================================
# PDF Processing
pymupdf==1.23.21  # Fast PDF parsing (PyMuPDF)
pdfplumber==0.10.3  # Table extraction
pypdf==4.0.1  # Fallback parser

# DOCX Processing
python-docx==1.1.0
docx2txt==0.8

# PPTX Processing  
python-pptx==0.6.23

# Image & OCR
pillow==10.2.0
pytesseract==0.3.10  # OCR for scanned documents
pdf2image==1.17.0

# Universal document parser (fallback)
unstructured[local-inference]==0.12.4
unstructured-inference==0.7.23

# ============================================================================
# Embeddings & ML
# ============================================================================
numpy==1.26.3
scipy==1.12.0
scikit-learn==1.4.0
transformers==4.37.2
torch==2.2.0  # For model inference
einops==0.7.0  # Einstein Operations for tensor manipulation (required for vision models)
opencv-python==4.9.0.80  # Computer vision library for image processing

# ============================================================================
# Caching & Storage
# ============================================================================
redis==5.0.1
aioredis==2.0.1
hiredis==2.3.2  # Fast Redis protocol parser

# Object storage
boto3==1.34.34  # AWS S3
minio==7.2.3  # MinIO (S3-compatible)

# ============================================================================
# LLM Providers (Tier 3 - Optional)
# ============================================================================
openai==1.12.0
anthropic==0.18.0
google-generativeai==0.3.2

# Ollama for local LLMs
ollama>=0.1.7

# ============================================================================
# Data Processing
# ============================================================================
pandas==2.2.0  # Table processing
chardet==5.2.0  # Character encoding detection
beautifulsoup4==4.12.3  # HTML parsing
lxml==5.1.0  # XML processing
markdownify==0.11.6  # HTML to Markdown
markdown==3.5.2  # Markdown processing

# ============================================================================
# CLI & UI
# ============================================================================
typer[all]==0.9.0  # Modern CLI
rich==13.7.0  # Beautiful terminal output
tqdm==4.66.1  # Progress bars
click==8.1.7

# ============================================================================
# API & Web
# ============================================================================
websockets==12.0  # WebSocket support
python-socketio==5.11.0  # Socket.IO
sse-starlette==2.0.0  # Server-Sent Events
jinja2==3.1.3  # Templates

# ============================================================================
# Logging
# ============================================================================
structlog==24.1.0  # Structured logging

# ============================================================================
# Testing & Quality
# ============================================================================
pytest>=7.4.0,<8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-benchmark==4.0.0
hypothesis==6.98.0  # Property-based testing

# Code quality
black==24.1.1  # Formatter
ruff==0.2.1  # Fast linter
mypy==1.8.0  # Type checker
pre-commit==3.6.0

# ============================================================================
# Utilities
# ============================================================================
python-dotenv==1.0.1  # Environment variables
pyyaml==6.0.1  # Configuration
toml==0.10.2
python-dateutil==2.8.2
pytz==2024.1

# Hashing & Crypto
xxhash==3.4.1  # Fast hashing for fingerprints

# Note: Database dependencies removed - using Redis for all persistence needs

# ============================================================================
# Development Tools
# ============================================================================
ipython==8.20.0
jupyter==1.0.0
ipdb==0.13.13  # Debugger
watchdog==4.0.0  # File watching for auto-reload

# ============================================================================
# Documentation
# ============================================================================
mkdocs==1.5.3
mkdocs-material==9.5.6
mkdocstrings[python]==0.24.0

# ============================================================================
# Container & Deployment
# ============================================================================
gunicorn==21.2.0  # WSGI server
supervisor==4.2.5  # Process control
